Revert "Refactor : operation and Directory struct"
