+++ aes/aes_test.go
@@ -377,6 +377,32 @@ func TestKeyExpansion(t *testing.T) {
 	}
 }
 
+func TestMul(t *testing.T) {
+	inputs := [][]byte{
+		[]byte{0x57, 0x13},
+		[]byte{0x57, 0x02},
+		[]byte{0x57, 0x04},
+		[]byte{0x57, 0x08},
+		[]byte{0x57, 0x10},
+		[]byte{0x57, 0x83},
+	}
+	expected := []byte{
+		0xfe,
+		0xae,
+		0x47,
+		0x8e,
+		0x07,
+		0xc1,
+	}
+
+	for i, input := range inputs {
+		result := mul(input[0], input[1])
+		if result != expected[i] {
+			t.Errorf("[TestMul] case %d failed: result != expected : '%v' != '%v'", i, result, expected[i])
+		}
+	}
+}
+
 func TestRotWord(t *testing.T) {
 	inputs := [][]byte{
 		[]byte{0x09, 0xcf, 0x4f, 0x3c},
@@ -437,38 +463,12 @@ func TestShiftRows(t *testing.T) {
 	Nb = BlockSize128
 
 	inputs := [][]byte{
-		[]byte{
-			0x63, 0xca, 0xb7, 0x04,
-			0x09, 0x53, 0xd0, 0x51,
-			0xcd, 0x60, 0xe0, 0xe7,
-			0xba, 0x70, 0xe1, 0x8c,
-		},
-		[]byte{
-			0xd4, 0x27, 0x11, 0xae,
-			0xe0, 0xbf, 0x98, 0xf1,
-			0xb8, 0xb4, 0x5d, 0xe5,
-			0x1e, 0x41, 0x52, 0x30,
-		},
-		[]byte{
-			0xd4, 0x27, 0x11, 0xae,
-			0xe0, 0xbf, 0x98, 0xf1,
-			0xb8, 0xb4, 0x5d, 0xe5,
-			0x1e, 0x41, 0x52, 0x30,
-		},
+		[]byte{0x63, 0xca, 0xb7, 0x04, 0x09, 0x53, 0xd0, 0x51, 0xcd, 0x60, 0xe0, 0xe7, 0xba, 0x70, 0xe1, 0x8c},
+		[]byte{0xd4, 0x27, 0x11, 0xae, 0xe0, 0xbf, 0x98, 0xf1, 0xb8, 0xb4, 0x5d, 0xe5, 0x1e, 0x41, 0x52, 0x30},
 	}
 	expected := [][]byte{
-		[]byte{
-			0x63, 0x53, 0xe0, 0x8c,
-			0x09, 0x60, 0xe1, 0x04,
-			0xcd, 0x70, 0xb7, 0x51,
-			0xba, 0xca, 0xd0, 0xe7,
-		},
-		[]byte{
-			0xd4, 0xbf, 0x5d, 0x30,
-			0xe0, 0xb4, 0x52, 0xae,
-			0xb8, 0x41, 0x11, 0xf1,
-			0x1e, 0x27, 0x98, 0xe5,
-		},
+		[]byte{0x63, 0x53, 0xe0, 0x8c, 0x09, 0x60, 0xe1, 0x04, 0xcd, 0x70, 0xb7, 0x51, 0xba, 0xca, 0xd0, 0xe7},
+		[]byte{0xd4, 0xbf, 0x5d, 0x30, 0xe0, 0xb4, 0x52, 0xae, 0xb8, 0x41, 0x11, 0xf1, 0x1e, 0x27, 0x98, 0xe5},
 	}
 
 	for i, input := range inputs {
@@ -527,32 +527,12 @@ func TestInvShiftRows(t *testing.T) {
 	Nb = BlockSize128
 
 	inputs := [][]byte{
-		[]byte{
-			0x7a, 0xd5, 0xfd, 0xa7,
-			0x89, 0xef, 0x4e, 0x27,
-			0x2b, 0xca, 0x10, 0x0b,
-			0x3d, 0x9f, 0xf5, 0x9f,
-		},
-		[]byte{
-			0x54, 0xd9, 0x90, 0xa1,
-			0x6b, 0xa0, 0x9a, 0xb5,
-			0x96, 0xbb, 0xf4, 0x0e,
-			0xa1, 0x11, 0x70, 0x2f,
-		},
+		[]byte{0x7a, 0xd5, 0xfd, 0xa7, 0x89, 0xef, 0x4e, 0x27, 0x2b, 0xca, 0x10, 0x0b, 0x3d, 0x9f, 0xf5, 0x9f},
+		[]byte{0x54, 0xd9, 0x90, 0xa1, 0x6b, 0xa0, 0x9a, 0xb5, 0x96, 0xbb, 0xf4, 0x0e, 0xa1, 0x11, 0x70, 0x2f},
 	}
 	expected := [][]byte{
-		[]byte{
-			0x7a, 0x9f, 0x10, 0x27,
-			0x89, 0xd5, 0xf5, 0x0b,
-			0x2b, 0xef, 0xfd, 0x9f,
-			0x3d, 0xca, 0x4e, 0xa7,
-		},
-		[]byte{
-			0x54, 0x11, 0xf4, 0xb5,
-			0x6b, 0xd9, 0x70, 0x0e,
-			0x96, 0xa0, 0x90, 0x2f,
-			0xa1, 0xbb, 0x9a, 0xa1,
-		},
+		[]byte{0x7a, 0x9f, 0x10, 0x27, 0x89, 0xd5, 0xf5, 0x0b, 0x2b, 0xef, 0xfd, 0x9f, 0x3d, 0xca, 0x4e, 0xa7},
+		[]byte{0x54, 0x11, 0xf4, 0xb5, 0x6b, 0xd9, 0x70, 0x0e, 0x96, 0xa0, 0x90, 0x2f, 0xa1, 0xbb, 0x9a, 0xa1},
 	}
 
 	for i, input := range inputs {
