+++ aes/utils.go
@@ -4,25 +4,35 @@ import (
 	"fmt"
 )
 
-func xtime(n byte) byte {
-	p := int(n) << 1
-	if p&0x100 != 0 {
-		p ^= poly
+func mul(num1, num2 byte) byte {
+	switch num2 {
+	case 0:
+		return 0
+	case 1:
+		return num1
+	case 2:
+		return mul2(num1)
 	}
-	return byte(p)
+
+	remains := int(num2)
+	result := num1
+	i := 2
+	for ; i <= remains; i *= 2 {
+		result = mul2(result)
+	}
+	remains -= i / 2
+	if remains > 0 {
+		result ^= mul(num1, byte(remains))
+	}
+	return result
 }
 
-func mul(n, p byte) byte {
-	switch {
-	case p == 0:
-		return 0
-	case p%2 == 0:
-		t := mul(n, p/2)
-		return xtime(t)
-	case p%2 == 1:
-		return n ^ mul(n, p-1)
+func mul2(num byte) byte {
+	tmp := int(num) << 1 // multiplied by 2
+	if tmp&0x100 != 0 {  //if 0x100 is set
+		tmp ^= 0x11b // mod by 0x11b (x^8 + x^4 + x^3 + x + 1)
 	}
-	return n
+	return byte(tmp)
 }
 
 func printBytes(bytes []byte) {
